refactor: simplify GetPagedAsync call in ArticleService
The code change in the `ArticleService.cs` file within the
`home_wiki_backend.BL.Services` namespace involves modifying the way
the `GetPagedAsync` method is called. Specifically, the
`filter.GetPredicate()` and `filter.GetOrderBy()` parameters have been
replaced with a new instance of `ArticleForFilterSpecification` that
takes `filter` as an argument.

This modification simplifies the call to `GetPagedAsync` by
encapsulating the filter logic within the `ArticleForFilterSpecification`
class, potentially improving readability and maintainability of the code.